Use 32 bit values for stream slots
We observed some strange streaming behavior when the stream slot
assignment wraps around the maximum value. By increasing the number
of slots to a value that we won't hit for a long time we effectively
work around this problem.

+++ libcaf_core/caf/stream_slot.hpp
@@ -13,7 +13,7 @@ namespace caf {
 
 /// Identifies a single stream path in the same way a TCP port identifies a
 /// connection over IP.
-using stream_slot = uint16_t;
+using stream_slot = uint32_t;
 
 /// Identifies an invalid slot.
 constexpr stream_slot invalid_stream_slot = 0;
